* merge internal code; txn opt, cost index select

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* merge internal code; txn opt, cost index select;performace for seek and prepare stmt; add arrow 0.17.1

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

* fix compile

+++ src/logical_plan/query_context.cpp
@@ -18,10 +18,13 @@
 namespace baikaldb {
 DEFINE_bool(default_2pc, false, "default enable/disable 2pc for autocommit queries");
 QueryContext::~QueryContext() {
-    ExecNode::destroy_tree(root);
+    if (need_destroy_tree) {
+        ExecNode::destroy_tree(root);
+    }
 }
 
 int QueryContext::create_plan_tree() {
+    need_destroy_tree = true;
     return ExecNode::create_tree(plan, &root);
 }
